Esta máquina ya sirve.
