--- Practica3/src/main/java/mx/unam/ciencias/alf/practica3/Practica3.java
@@ -1,20 +1,0 @@
-package mx.unam.ciencias.alf.practica3;
-
-import com.google.gson.Gson;
-
-/**
- * @author Berenice Calvario Gonzalez
- * @author Jesús Alberto Reyes Gutiérrez
- * @param args
- * @throws java.io.IOException
- */
-
-
-public class Practica3 {
-
-    public static void main(String[] args) {
-        //Gson g = new Gson();
-
-        System.out.println("Hola");
-    }
-}
